1
更新了第三版的超声波代码

+++ Match_code/mcu超声波实验/Hardware/Beep.h
@@ -4,5 +4,5 @@
 #include "sys.h"
 void Beep_Init();
 void Set_Beep_compare(int Compare);
-int Dist_tocompare(float dist);
+int Dist_tocompare(int dist);
 #endif
